Rename file knora-api-params.ts to knora-request.ts
