+++ src/client/app/shared/utilities/knora-request.ts
